+++ app/android/app/src/main/kotlin/me/jxl/kiosk_satellite/KioskLock.kt
@@ -173,6 +173,9 @@ class KioskLock(private val activity: Activity, messenger: BinaryMessenger) {
                     call.argument<String>("cutout")?.let {
                         CutoutLayout.apply(activity, it)
                     }
+                    call.argument<String>("orientation")?.let {
+                        ScreenOrientation.apply(activity, it)
+                    }
                     result.success(null)
                 }
                 "navCapture" -> {